+++ src/merchantPurchase/merchantPurchase.php
@@ -21,6 +21,8 @@ class MerchantPurchase extends Resource
     public $holderName;
     public $holderEmail;
     public $holderPhone;
+    public $holderId;
+    public $softDescriptor;
     public $billingCountryCode;
     public $billingCity;
     public $billingStateCode;
@@ -57,6 +59,8 @@ class MerchantPurchase extends Resource
         $this->holderName = Checks::checkParam($params, "holderName");
         $this->holderEmail = Checks::checkParam($params, "holderEmail");
         $this->holderPhone = Checks::checkParam($params, "holderPhone");
+        $this->holderId = Checks::checkParam($params, "holderId");
+        $this->softDescriptor = Checks::checkParam($params, "softDescriptor");
         $this->fundingType = Checks::checkParam($params, "fundingType");
         $this->billingCountryCode = Checks::checkParam($params, "billingCountryCode");
         $this->billingCity = Checks::checkParam($params, "billingCity");