+++ src/merchantSession/purchase.php
@@ -16,6 +16,8 @@ class Purchase extends Resource
     public $holderName;
     public $holderEmail;
     public $holderPhone;
+    public $holderId;
+    public $softDescriptor;
     public $fundingType;
     public $billingCountryCode;
     public $billingCity;
@@ -54,6 +56,8 @@ class Purchase extends Resource
         $this->holderName = Checks::checkParam($params, "holderName");
         $this->holderEmail = Checks::checkParam($params, "holderEmail");
         $this->holderPhone = Checks::checkParam($params, "holderPhone");
+        $this->holderId = Checks::checkParam($params, "holderId");
+        $this->softDescriptor = Checks::checkParam($params, "softDescriptor");
         $this->fundingType = Checks::checkParam($params, "fundingType");
         $this->billingCountryCode = Checks::checkParam($params, "billingCountryCode");
         $this->billingCity = Checks::checkParam($params, "billingCity");